Fix ToT fails to build during google-benchmark step
We do not run any benchmarks, so there is no need to build them.
However, for some reason, setting the LIBCXX_INCLUDE_BENCHMARKS option
through CMake invocation takes no effect.

Workaround this by changing the source to be OFF by default, to unblock
the ToT testing. We need to find a proper way to fix this in the long
term...

Test: toolchain/llvm_android/build.py --build-llvm-next --llvm-rev=main --no-build=windows,lldb
Bug: 386218327
Change-Id: If8d777112edc2af7cdbeb4b62ffe25141423345c