Copy profiles if they are set via environment variables
Copy PGO and BOLT profiles to DIST_DIR if they are passed via
environment variables.  This typically only happens on CI builds.  In
aosp-llvm-toolchain, the copy was part of the GCL build command and was
lost when the new branches were set up.  Do the copy in our build
scripts and keep the new GCL configs simple.

Also clean up extract_pgo_profile - the `pgo` parameter, derived from
the `--pgo` boolean argument is unlikely to be a string.

Test: Run build.py with LLVM_{PGO,BOLT}_PROFILE and verify profiles are
      copied.
Change-Id: I0bf69e38fda886fa258768c7e2e831aa187b0d14