Remove fallback mode from test_compiler
Bug: http://b/228206914

test_compiler and the compiler wrapper support a mode where it fallsback
to the old compiler in case of a failure.  This mode has been off for a
while, and moreover does not work with RBE and also has potential
problems with LTO.  Remove this functionality from test_compiler.py.  We
could also remove this from the wrapper if it is only used in Android.

Test: presubmit
Change-Id: I76fd19b884707f2c30e0b793cdf6db8b4a115edd