Build comparison: add filtering and aggregation
Added a method to the build configuration class to filter build traces
based on program arguments.

Added a method to aggregate arbitrary dataframes representing a sample,
and produce mean values with desired confidence intervals. This will be
used to aggregate multiple trace files under one build configuration
into a single dataframe before comparisons.

Added a method to merge two dataframes and return the top K entries
with the largest relative differences in a given column name. This will
be used to find the build units of interest to the user.

Bug: 42304181
Change-Id: Iec4ee91e0001aaca5346484801f688646a859d75