Build comparison: add bar chart comparison
Added a method to compare two build configurations and display a bar
chart showing the targets with the largest relative differences in build
times.

Added a method to merge two data frames on an arbitrary key and return
the k rows with the greatest relative difference between two desired
values.

Bug: 42304181
Change-Id: Ie3922c12d85664889673a5dccf785832defb5fbe