Stop generating symlinks for prebuilts.
These symlinks that use legacy names with the architecture embedded in
the name itself, and are no longer used. Also remove code that verifies
the generated symlinks.

Because of an inconsistency between the value given to --march (i686)
and the triple name associated with it (i386), just create symlinks from
i686 to i386 until the inconsistency is addresed.

Bug: 245395722
Test: presubmits
Change-Id: I324d541151a23509c3269245e5e5500dd20f5839