Revert std::map commit temporarily to fix protobuf in Kokoro
https://github.com/llvm/llvm-project/pull/161485 is causing issues
with outdated protobuf in AOSP. The internal protobuf does not have this
issue anymore but it is difficult to just take a diff and make a kokoro
patch since protobuf updates affect other projects. The temporary
solution is to revert the LLVM CL to make ToT Kokoro AOSP builds green
again while I work on a fix for the operator (std::less) issue in
protobuf. Once we have a better workaround, we can delete this local
patch.

Bug: 450376246
Test: toolchain/llvm_android/build.py --build-llvm-next --llvm-rev
573b624899d4e03a5249207ae5bb2a0f4fabb28b --no-build=windows,lldb
--no-lto --skip-tests
Test: m with ToT Build

Change-Id: I1f76ac950628b54aa108b1aa45066961324ce31c