Apply ToT patches when building ToT only
ToT patches are only meant to applied to keep ToT builds green until
a permanent solution is found. We created a json for ToT patches and
apply them only when `--build-llvm-next` is passed to the build script.
This will avoid confusing between downstream patches needed for
compiler updates and Kokoro.

Test: toolchain/llvm_android/build.py --build-llvm-next --llvm-rev
0198e8f068460f292477a6797de31aa4bac736f5 --no-build=windows,lldb
--no-lto
Test: toolchain/llvm_android/build.py --no-build=windows,lldb --no-lto

Change-Id: I512440ae5a482bdafe2204a64d966974854ed20c