Fix path in Kokoro patch
I create this patch relative to the internal git repo after testing but
the kokoro patches are relative to the top of Android's tree. This
caused the patch to be applied to the wrong place and fail. I have now
amde it relative to the top of Android's directory and tested it with
patch.

Bug: 466109356
Test: patch -p1 <
toolchain/llvm_android/kokoro/tot-patches/0003-trusty-security-build-removal.patch

Change-Id: Ibb43abb872f057f0b83588b2111a918c74e8a99f